backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.350
langchain-google-genai==0.0.5
chromadb==0.4.15
python-dotenv==1.0.0
pydantic==2.5.0
